Add test checking that invalid votes aren't counted
The circuit should not allow a prover to generate a valid proof without
real votes (real votes = voteValue + valid signature), as then, it would
be possible to present a proof that would pass verifications, claiming a
result that has no real votes. Meaning that a prover could provide a
valid zk-proof claiming a result without real votes behind it, and the
verification would accept it.

This commit adds a test that fails if the circuit allows the described
behaviour, counting invalid votes, simulating a malicious prover.